[#44] Refactor: 중복되는 타입 코드 분리
